+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 Version 1.4-SNAPSHOT
 -------------
 
+ADDED:
+
+- new class PrefixFileFilter created that can find files with a given prefix recursive
+
+CHANGED:
+
+- update of gradle to new version 6.8
+- update of com.github.ben-manes.versions.gradle.plugin to new version 0.36.0
+- update of test dependency silly-collections in new version 8.3
+- all FileFilter and FilenameFilter classes have a recursive flag for recursive search
+
 Version 1.3
 -------------